+++ domain/primal/src/commonMain/kotlin/net/primal/domain/upload/UploadResult.kt
@@ -2,6 +2,6 @@ package net.primal.domain.upload
 
 data class UploadResult(
     val remoteUrl: String,
-    val originalFileSize: Int,
+    val originalFileSize: Long,
     val originalHash: String,
 )
